packages/hooks/src/useSetState/index.ts
import { useCallback, useState } from 'react';
import { isFunction } from '../utils';

type SetState<S extends Record<string, any>> = <K extends keyof S>(
  state: Pick<S, K> | null | ((prevState: Readonly<S>) => Pick<S, K> | S | null),
) => void;

type ResetState = () => void;

const useSetState = <S extends Record<string, any>>(
  initialState: S | (() => S),
): [S, SetState<S>, ResetState] => {
  const [state, setState] = useState<S>(initialState);

  const setMergeState = useCallback((patch) => {
    setState((prevState) => {
      const newState = isFunction(patch) ? patch(prevState) : patch;
      return newState ? { ...prevState, ...newState } : prevState;
    });
  }, []);

  const resetState = useCallback(() => {
    setState(initialState);
  }, []);

  return [state, setMergeState, resetState];
};

export default useSetState;
